+++ src/dev/sr_4021.h
@@ -11,7 +11,7 @@ namespace daisy
  ** @ingroup shiftregister
  **
  ** CD4021B-Q1: CMOS 8-STAGE STATIC SHIFT REGISTER
- ** 
+ **
  ** Supply Voltage: 3V to 18V
  ** Clock Freq: 3MHz at 5V (less at 3v3) -> 8.5MHz at 15V
  ** Pin Descriptions:
@@ -21,14 +21,14 @@ namespace daisy
  ** - P/!S               - 9
  ** - Q[6-8]             - 2, 12, 3
  **
- ** Driver has support for daisy chaining and running up to 2 same-sized 
- ** chains in parallel from a single set of clk/latch pins to reduce 
+ ** Driver has support for daisy chaining and running up to 2 same-sized
+ ** chains in parallel from a single set of clk/latch pins to reduce
  ** pin/code overhead when using multiple devices.
  **
- ** When dealing with multiple parallel/daisy-chained devices the 
+ ** When dealing with multiple parallel/daisy-chained devices the
  ** states of all inputs will be filled in the following order (example uses two chained and two parallel):
  ** data[chain0,parallel0], data[chain1,parallel0], data[chain0,parallel1], data[chain1,parallel1];
- ** 
+ **
  ** When combining multiple daisy chained and parallel devices the number of devices chained should match
  ** for each parallel device chain.
  **
@@ -40,9 +40,16 @@ class ShiftRegister4021
     /** Configuration Structure for handling the pin setting of the device */
     struct Config
     {
-        dsy_gpio_pin clk;   /**< Clock pin to attach to pin 10 of device(s) */
-        dsy_gpio_pin latch; /**< Latch pin to attach to pin 9 of device(s) */
-        dsy_gpio_pin data[num_parallel]; /**< Data Pin(s) */
+        Pin clk;   /**< Clock pin to attach to pin 10 of device(s) */
+        Pin latch; /**< Latch pin to attach to pin 9 of device(s) */
+        Pin data[num_parallel]; /**< Data Pin(s) */
+
+        /**
+         * DelayTicks between these actions helps to ensure stable timing
+         * for the connected device(s).
+         * Each tick is approx. 4.16ns at CPUFreq 480MHz
+         */
+        uint32_t delay_ticks = 10;
     };
 
     ShiftRegister4021() {}
@@ -53,20 +60,11 @@ class ShiftRegister4021
     {
         config_ = cfg;
         // Init GPIO
-        clk_.mode = DSY_GPIO_MODE_OUTPUT_PP;
-        clk_.pull = DSY_GPIO_NOPULL;
-        clk_.pin  = cfg.clk;
-        dsy_gpio_init(&clk_);
-        latch_.mode = DSY_GPIO_MODE_OUTPUT_PP;
-        latch_.pull = DSY_GPIO_NOPULL;
-        latch_.pin  = cfg.latch;
-        dsy_gpio_init(&latch_);
+        clk_.Init(cfg.clk, GPIO::Mode::OUTPUT);
+        latch_.Init(cfg.latch, GPIO::Mode::OUTPUT);
         for(size_t i = 0; i < num_parallel; i++)
         {
-            data_[i].mode = DSY_GPIO_MODE_INPUT;
-            data_[i].pull = DSY_GPIO_NOPULL;
-            data_[i].pin  = cfg.data[i];
-            dsy_gpio_init(&data_[i]);
+            data_[i].Init(cfg.data[i], GPIO::Mode::INPUT);
         }
         // Init States
         for(size_t i = 0; i < kTotalStates; i++)
@@ -78,30 +76,31 @@ class ShiftRegister4021
     /** Reads the states of all pins on the connected device(s) */
     void Update()
     {
-        dsy_gpio_write(&clk_, 0);
-        dsy_gpio_write(&latch_, 1);
-        System::DelayTicks(1);
-        dsy_gpio_write(&latch_, 0);
+        uint32_t del_ticks = config_.delay_ticks;
+        clk_.Write(false);
+        latch_.Write(true);
+        System::DelayTicks(del_ticks);
+        latch_.Write(false);
         uint32_t idx;
         for(size_t i = 0; i < 8 * num_daisychained; i++)
         {
-            dsy_gpio_write(&clk_, 0);
-            System::DelayTicks(1);
+            clk_.Write(false);
+            System::DelayTicks(del_ticks);
             for(size_t j = 0; j < num_parallel; j++)
             {
                 idx = (8 * num_daisychained - 1) - i;
                 idx += (8 * num_daisychained * j);
-                states_[idx] = dsy_gpio_read(&data_[j]);
+                states_[idx] = data_[j].Read();
             }
-            dsy_gpio_write(&clk_, 1);
-            System::DelayTicks(1);
+            clk_.Write(true);
+            System::DelayTicks(del_ticks);
         }
     }
 
-    /** returns the last read state of the input at the index. 
+    /** returns the last read state of the input at the index.
      ** true indicates the pin is held HIGH.
-     ** 
-     ** See above for the layout of data when using multiple 
+     **
+     ** See above for the layout of data when using multiple
      ** devices in series or parallel.
      ***/
     inline bool State(int index) const { return states_[index]; }
@@ -112,9 +111,9 @@ class ShiftRegister4021
     static constexpr int kTotalStates = 8 * num_daisychained * num_parallel;
     Config               config_;
     bool                 states_[kTotalStates];
-    dsy_gpio             clk_;
-    dsy_gpio             latch_;
-    dsy_gpio             data_[num_parallel];
+    GPIO                 clk_;
+    GPIO                 latch_;
+    GPIO                 data_[num_parallel];
 };
 
 } // namespace daisy